+++ CHANGELOG.md
@@ -72,6 +72,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Cursor being hidden after reaching cursor blinking timeout
 - Message bar content getting stuck after closing with multiple messages on Wayland
 - Vi cursor position not redrawn on PageUp/PageDown without scrollback
+- Cursor not updating when blinking and viewport is scrolled
 
 ### Removed